feat: remove working

+++ src/cli.rs
@@ -13,18 +13,29 @@ pub fn list() -> Result<(), String> {
     let app_paths = utils::get_app_paths();
     let gitconfig_data = git::get_gitconfig_data(&app_paths.gitconfig_file);
     let profile_dirs: Vec<String> = utils::get_dirs(&app_paths.data_dir).unwrap_or_else(|_| vec![]);
+
     let all_ssh_files: Vec<String> =
         utils::get_files(&app_paths.ssh_dir).unwrap_or_else(|_| vec![]);
-    let ssh_files: Vec<String> = all_ssh_files
+    let mut ssh_files: Vec<String> = all_ssh_files
         .into_iter()
         .filter(|filename| SSH_FILES.contains(&filename.as_str()))
         .collect();
 
-    let mut current_profile_sum_paths: Vec<PathBuf> = vec![];
-    for ssh_file in &ssh_files {
-        current_profile_sum_paths.push(app_paths.ssh_dir.join(ssh_file));
+    if gitconfig_data.file_exists {
+        ssh_files.push(GITCONFIG_FILE.to_string());
     }
-    current_profile_sum_paths.push(app_paths.gitconfig_file);
+
+    ssh_files.sort();
+
+    let mut current_profile_sum_paths: Vec<PathBuf> = vec![];
+
+    ssh_files.iter().for_each(|filename| {
+        if filename.as_str() == GITCONFIG_FILE {
+            current_profile_sum_paths.push(app_paths.gitconfig_file.clone());
+        } else {
+            current_profile_sum_paths.push(app_paths.ssh_dir.join(&filename));
+        }
+    });
 
     let current_hash =
         utils::get_profile_hash(&current_profile_sum_paths).unwrap_or_else(|_| String::from("---"));
@@ -34,20 +45,27 @@ pub fn list() -> Result<(), String> {
 
     for profile_dir in profile_dirs {
         let mut prefix: &str = " ";
-        let mut profile_sum_paths: Vec<PathBuf> = vec![];
         let profile_path = app_paths.data_dir.join(&profile_dir);
-        let all_profile_files = utils::get_files(&profile_path)?;
-        all_profile_files
+        let all_profile_files: Vec<String> = utils::get_files(&profile_path)?;
+        let mut filtered_files: Vec<String> = all_profile_files
             .into_iter()
             .filter(|filename| {
                 SSH_FILES.contains(&filename.as_str()) || filename.as_str() == GITCONFIG_FILE
             })
+            .collect();
+        filtered_files.sort();
+
+        let mut profile_sum_paths: Vec<PathBuf> = vec![];
+
+        filtered_files
+            .into_iter()
             .for_each(|filename| profile_sum_paths.push(profile_path.join(filename)));
+
         let profile_hash = utils::get_profile_hash(&profile_sum_paths).expect(READING_FILES_ERR);
 
         if profile_hash == current_hash {
             prefix = "*";
-            current_profile_name.push_str(&profile_dir);
+            current_profile_name = profile_dir.clone();
         }
 
         println!("{} {}", prefix, profile_dir);
@@ -55,12 +73,16 @@ pub fn list() -> Result<(), String> {
     println!("");
 
     if current_profile_name.is_empty() {
-        println!("--- current profile not saved ---");
+        println!("--- no profile has been saved for the current files ---");
     } else {
         println!("(current = {})", current_profile_name);
     }
-    println!("  gc_name: {:?}", gitconfig_data.name);
-    println!("  gc_mail: {:?}", gitconfig_data.email);
+
+    if gitconfig_data.file_exists {
+        println!("  gitconfig_name: {:?}", gitconfig_data.name);
+        println!("  gitconfig_mail: {:?}", gitconfig_data.email);
+    }
+
     println!("  files ({:?}):", ssh_files.len());
     for filename in ssh_files {
         println!("    {}", filename);
@@ -112,13 +134,8 @@ pub fn save(profile_name: &str) -> Result<(), String> {
     }
 
     let app_paths = utils::get_app_paths();
+    let gitconfig_data = git::get_gitconfig_data(&app_paths.gitconfig_file);
     let profile_path = app_paths.data_dir.join(profile_name);
-
-    if let Err(err) = fs::remove_dir_all(&profile_path) {
-        if err.kind() != ErrorKind::NotFound {
-            return Err(REMOVING_DIR_ERR.to_string());
-        }
-    }
 
     let all_ssh_files: Vec<String> =
         utils::get_files(&app_paths.ssh_dir).unwrap_or_else(|_| vec![]);
@@ -127,22 +144,65 @@ pub fn save(profile_name: &str) -> Result<(), String> {
         .filter(|filename| SSH_FILES.contains(&filename.as_str()))
         .collect();
 
-    if let Err(_) = utils::copy_file(
-        &app_paths.gitconfig_file,
-        &profile_path.join(GITCONFIG_FILE),
-    ) {
-        eprintln!("Error: Could not copy file: {:?}", GITCONFIG_FILE);
+    if !gitconfig_data.file_exists && ssh_files.len() == 0 {
+        return Err("No files found to save for this profile.".to_string());
     }
+
+    if let Err(err) = fs::remove_dir_all(&profile_path) {
+        if err.kind() != ErrorKind::NotFound {
+            return Err(REMOVING_DIR_ERR.to_string());
+        }
+    }
+
+    if gitconfig_data.file_exists {
+        if let Err(_) = utils::copy_file(
+            &app_paths.gitconfig_file,
+            &profile_path.join(GITCONFIG_FILE),
+        ) {
+            return Err(format!("Error: Could not copy file: {:?}", GITCONFIG_FILE));
+        } else {
+            println!("File copied!: {}", GITCONFIG_FILE)
+        }
+    }
+
     for ssh_file in &ssh_files {
         if let Err(_) = utils::copy_file(
             &app_paths.ssh_dir.join(ssh_file),
             &profile_path.join(ssh_file),
         ) {
-            eprintln!("Error: Could not copy file: {:?}", ssh_file);
+            return Err(format!("Error: Could not copy file: {:?}", ssh_file));
+        } else {
+            println!("File copied!: {}", ssh_file)
         }
     }
 
     println!("Saved profile {:?} successfully!", profile_name);
+    Ok(())
+}
+
+pub fn remove(profile_name: &str) -> Result<(), String> {
+    if profile_name.is_empty() {
+        let lines = [
+            "Profile name cannot be empty.",
+            "Example: xks remove alex_github",
+        ];
+        let msg = lines.join("\n");
+        return Err(msg);
+    }
+
+    let app_paths = utils::get_app_paths();
+    let profile_path = app_paths.data_dir.join(profile_name);
+
+    if let Err(err) = fs::remove_dir_all(&profile_path) {
+        if err.kind() != ErrorKind::NotFound {
+            return Err(REMOVING_DIR_ERR.to_string());
+        } else {
+            println!("Profile not found. Nothing to remove.");
+            return Ok(());
+        }
+    }
+
+    println!("Removed profile {:?} successfully!", profile_name);
     Ok(())
 }
